+++ Film-in/Common/Data/DTO/SearchPerson/SearchPersonResposeDTO+Mapping.swift
@@ -1,5 +1,5 @@
 //
-//  SearchMovieResponseDTO.swift
+//  SearchPersonResposeDTO.swift
 //  Film-in
 //
 //  Created by Minjae Kim on 9/23/24.
@@ -7,9 +7,9 @@
 
 import Foundation
 
-struct SearchMovieResponseDTO: Decodable {
+struct SearchPersonResposeDTO: Decodable {
     let page: Int
-    let results: [TMDBMovieResponseDTO]
+    let results: [TMDBPersonResponseDTO]
     let totalPage: Int
     
     enum CodingKeys: String, CodingKey {
